edited destroy action in CocktailsController, displays a message when a cocktail cannot be destroyed

+++ wood-tavern-backend/app/controllers/cocktails_controller.rb
@@ -40,7 +40,11 @@ class CocktailsController < ApplicationController
 
   # DELETE /cocktails/1
   def destroy
-    cocktail.destroy
+    cocktail = Cocktail.find_by(id: params[:id])
+    if cocktail
+      cocktail.destroy
+    else
+      render json: { message: 'Cocktail not found' }
   end
 
   private
